+++ rl2/agents_v2/architectures/lstm.py
@@ -50,7 +50,7 @@ class LSTM(tc.nn.Module):
 
     def forward(
         self,
-        input_vec: tc.LongTensor,
+        input_vec: tc.FloatTensor,
         prev_state: tc.FloatTensor
     ) -> Tuple[tc.FloatTensor, tc.FloatTensor]:
         """
